feat(tracking): add timestamp to Measurement and create SeedMeasurementCommand

+++ src/main/java/com/novaperutech/veyra/platform/tracking/domain/model/valueobjects/ResidentId.java
@@ -3,7 +3,7 @@ package com.novaperutech.veyra.platform.tracking.domain.model.valueobjects;
 import jakarta.persistence.Embeddable;
 
 @Embeddable
-public record ResidentId(Integer residentId) {
+public record ResidentId(Long residentId) {
 
     public ResidentId {
         if (residentId == null || residentId <= 0) {
